Рефакторинг и мелкие правки в CLI
