+++ lib/dapp.rb
@@ -46,6 +46,7 @@ require 'dapp/error/config'
 require 'dapp/error/shellout'
 require 'dapp/exception/base'
 require 'dapp/cli'
+require 'dapp/cli/command/options/tag'
 require 'dapp/cli/command/base'
 require 'dapp/cli/command/update'
 require 'dapp/cli/command/slug'
