+++ lib/dapp/dimg/cli/command/dimg/push.rb
@@ -11,6 +11,8 @@ Usage:
 
 Options:
 BANNER
+        extend ::Dapp::CLI::Command::Options::Tag
+
         option :lock_timeout,
                long: '--lock-timeout TIMEOUT',
                description: 'Redefine resource locking timeout (in seconds)',
@@ -19,32 +21,6 @@ BANNER
         option :git_artifact_branch,
                long: '--git-artifact-branch BRANCH',
                description: 'Default branch to archive artifacts from'
-
-        option :tag,
-               long: '--tag TAG',
-               description: 'Add tag (can be used one or more times)',
-               default: [],
-               proc: proc { |v| composite_options(:tags) << v }
-
-        option :tag_branch,
-               long: '--tag-branch',
-               description: 'Tag by git branch',
-               boolean: true
-
-        option :tag_build_id,
-               long: '--tag-build-id',
-               description: 'Tag by CI build id',
-               boolean: true
-
-        option :tag_ci,
-               long: '--tag-ci',
-               description: 'Tag by CI branch and tag',
-               boolean: true
-
-        option :tag_commit,
-               long: '--tag-commit',
-               description: 'Tag by git commit',
-               boolean: true
 
         option :with_stages,
                long: '--with-stages',
