+++ lib/dapp/dimg/cli/command/dimg/stages/base.rb
@@ -3,8 +3,8 @@ module Dapp::Dimg::CLI
     class Dimg < ::Dapp::CLI
       class Stages < ::Dapp::CLI
         class Base < Base
-          def run_dapp_command(run_command, options: {}, log_running_time: true)
-            super(run_command, options: options.merge(verbose: true), log_running_time: log_running_time)
+          def run_dapp_command(run_command, options: {}, log_running_time: true, **extra_options)
+            super(run_command, options: options.merge(verbose: true), log_running_time: log_running_time, **extra_options)
           end
         end
       end
