+++ lib/dapp/kube/cli/command/kube/deploy.rb
@@ -8,6 +8,8 @@ Usage:
 
 Options:
 BANNER
+      extend ::Dapp::CLI::Command::Options::Tag
+
       option :namespace,
              long: '--namespace NAME',
              default: nil
@@ -17,32 +19,6 @@ BANNER
              description: "Custom tag (alias for --tag)",
              default: [],
              proc: proc { |v| composite_options(:image_versions) << v }
-
-      option :tag,
-             long: '--tag TAG',
-             description: 'Custom tag',
-             default: [],
-             proc: proc { |v| composite_options(:tags) << v }
-
-      option :tag_branch,
-             long: '--tag-branch',
-             description: 'Tag by git branch',
-             boolean: true
-
-      option :tag_build_id,
-             long: '--tag-build-id',
-             description: 'Tag by CI build id',
-             boolean: true
-
-      option :tag_ci,
-             long: '--tag-ci',
-             description: 'Tag by CI branch and tag',
-             boolean: true
-
-      option :tag_commit,
-             long: '--tag-commit',
-             description: 'Tag by git commit',
-             boolean: true
 
       option :tmp_dir_prefix,
              long: '--tmp-dir-prefix PREFIX',
