+++ lib/dapp/kube/cli/command/kube/secret_key_generate.rb
@@ -8,8 +8,8 @@ Usage:
 
 Options:
 BANNER
-      def run_dapp_command(run_method, options: {}, log_running_time: false)
-        super(run_method, options: options, log_running_time: log_running_time)
+      def run_dapp_command(run_method, options: {}, log_running_time: false, **extra_options)
+        super(run_method, options: options, log_running_time: log_running_time, **extra_options)
       end
     end
   end
